+++ modules/template/models/FileContent.php
@@ -115,4 +115,14 @@ use humhub\modules\file\models\File;
         ]);
     }
 
-}
+    public function canEdit($user = null): bool
+    {
+        return PagePermission::canEdit();
+    }
+
+    public function isEmpty(): bool
+    {
+        return !$this->hasFile();
+    }
+
+ }