+++ docs/CHANGELOG.md
@@ -7,6 +7,7 @@ Changelog
 - Fix #291: Fix broken URL in email footer
 - Fix #293: Initialize module content class
 - Enh #301: Tests for `next` version
+- Enh #318: Allow resetting of file/image content to default
 
 1.9.3 (June 13, 2023)
 ---------------------